Back of breadboard
The back of the breadboard is seen in this picture and it shows the power source for this circuit. It is a battery pack gathered from an remote control helicopter. It is a 4.5 voltage source which is good considering that the PIC18f4321 cannot handle high voltage and it makes a good portable battery pack.